backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-pinecone
pinecone-client
pytesseract
pillow
python-multipart